runs/scripts/evaluate.py
"""
Evaluation Script for Mirror Descent Experiments.

This script provides utilities for evaluating trained models and comparing
their implicit bias properties across different potential functions.
"""

import argparse
import yaml
import torch
import numpy as np
from pathlib import Path
from typing import Dict, Any, List

# TODO: Add imports for custom modules
# from core.models import DeepLinear, Linear
# from eval.metrics import compute_margin, compute_alignment, analyze_implicit_bias
# from eval.baselines import compare_baselines


def load_model(model_path: str, config: Dict[str, Any]) -> torch.nn.Module:
    """
    Load trained model from checkpoint.
    
    Args:
        model_path: Path to model checkpoint
        config: Configuration dictionary
        
    Returns:
        Loaded model
    """
    # TODO: Implement model loading
    pass


def run_comprehensive_evaluation(model: torch.nn.Module,
                               test_features: torch.Tensor,
                               test_labels: torch.Tensor,
                               potential_name: str) -> Dict[str, Any]:
    """
    Run comprehensive evaluation of model implicit bias.
    
    Args:
        model: Trained model
        test_features: Test features
        test_labels: Test labels
        potential_name: Name of potential function used
        
    Returns:
        Dictionary of evaluation results
    """
    # TODO: Implement comprehensive evaluation
    pass


def compare_potentials(results_dict: Dict[str, Dict[str, Any]]) -> Dict[str, Any]:
    """
    Compare results across different potential functions.
    
    Args:
        results_dict: Dictionary of results for each potential
        
    Returns:
        Comparison results
    """
    # TODO: Implement potential comparison
    pass


def generate_evaluation_report(results: Dict[str, Any], 
                             output_dir: str):
    """
    Generate comprehensive evaluation report.
    
    Args:
        results: Evaluation results
        output_dir: Output directory for report
    """
    # TODO: Implement report generation
    pass


def main():
    """Main evaluation function."""
    parser = argparse.ArgumentParser(description='Evaluate Mirror Descent models')
    parser.add_argument('--model_path', type=str, required=True,
                       help='Path to trained model')
    parser.add_argument('--config', type=str, default='configs/default.yaml',
                       help='Path to configuration file')
    parser.add_argument('--test_data', type=str, required=True,
                       help='Path to test data')
    parser.add_argument('--output_dir', type=str, default='evaluation_results',
                       help='Output directory for results')
    parser.add_argument('--baselines', action='store_true',
                       help='Run baseline comparisons')
    
    args = parser.parse_args()
    
    # TODO: Implement main evaluation logic
    print("Evaluation script placeholder - implement main evaluation logic")


if __name__ == '__main__':
    main()
