requirements.txt
# Core ML and Scientific Computing
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
scipy==1.11.4

# Machine Learning
scikit-learn==1.3.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Optional: JAX for alternative implementations
jax==0.4.23
jaxlib==0.4.23

# Configuration and Utilities
pyyaml==6.0.1
tqdm==4.66.1

# Development and Testing
pytest==7.4.3
pytest-cov==4.1.0

# Optional: Advanced Optimization
cvxpy==1.4.1
cvxopt==1.3.2

# Optional: LaTeX and Plotting
tikzplotlib==0.10.1

# Optional: Parallel Computing
joblib==1.3.2

# Optional: Data Processing
pandas==2.1.4
